some improvements to backup script

+++ 02_backup_to_S3/backup_to_S3.sh
@@ -2,8 +2,9 @@
 
 VERBOSE=false
 DRY_RUN=false
-DEV_PATH="None"
-S3_LINK="None"
+DEV_PATH=""
+S3_LINK=""
+RED=""
 
 # syslog logging func
 log_message() {
@@ -15,7 +16,7 @@ log_message() {
   fi
 }
 
-# Checking script's args
+# Checking script's args and validating input data
 args_setup() {
   while getopts ":vdp:s:" opt; do
     case "${opt}" in
@@ -27,22 +28,34 @@ args_setup() {
         DRY_RUN=true
         echo "Dry run mode on. Comands won't be executed, so you can check that script is running as it suppose to."
         ;;
-      p) 
-        DEV_PATH=$OPTARG
-        while [ -z "$$OPTARG" ]; do
-          read -p "Enter path to the logical volume. e.g. /dev/vg_some/lvol0\n Type: " DEV_PATH
-        done
+      p)
+        if [[ -b "$OPTARG" ]] && echo "$OPTARG" | grep "^/dev/*" > /dev/null; then
+          DEV_PATH=$OPTARG
+        else
+          echo "Check -p option. It isn't a block special file nor starts from '/dev/..'" "err"
+          exit 1
+        fi
         ;;
       s)
-        S3_LINK=$OPTARG
+        if echo "$OPTARG" | grep "^s3://" > /dev/null; then 
+          S3_LINK=$OPTARG
+        else
+          echo "Check -s option. Looks like the link isn't valid and doesn't start from 's3://'"
+          exit 1
+        fi
         ;;
       *)
-        echo "Invalid option"
-        echo "Usage: $0 [-v|-d=LV partition to backup|-s=s3 link to connect]"
-        exit 1
+        help_info
         ;;
     esac
   done
+}
+
+help_info() {
+  echo "Usage: $0 -p String (LV partition to backup) -s String (S3 link to connect)"
+  echo "        [ -v -- Verbose mode]"
+  echo "        [ -d -- Dry run ]"
+  exit 1
 }
 
 # VM cache flushing
@@ -67,7 +80,7 @@ create_snapshot() {
     echo "lvcreate --size 1G --snapshot --name backup_snapshot " "$DEV_PATH"
   else
     if ! lvcreate --size 1G --snapshot --name backup_snapshot $$DEV_PATH; then
-      log_message "Some problem has occur during creating snapshot" "error"
+      log_message "Some problem has occur during snapshot creation " "error"
       exit 1
     else
       log_message "Snapshot has been made succesfully"
@@ -77,21 +90,28 @@ create_snapshot() {
 
 # Making an tar.gz archive from LVM snapshot
 create_archive() {
-  tar -czf /backup/snapshot_backup.tar.gz /mnt/snapshot
+  if $DRY_RUN; then
+    echo "Making tar.gz archive"
+    echo "tar -czf /backup/snapshot_backup.tar.gz /mnt/snapshot"
+  else
+    tar -czf /backup/snapshot_backup.tar.gz /mnt/snapshot
+  fi
 }
 
 # Sending archive to S3 bucket
 upload_to_s3() {
-  aws s3 cp /backup/snapshot_backup.tar.gz s3://mybucket/
+  if $DRY_RUN; then
+    echo "Sending backup to S3 $S3_LINK" 
+    echo "aws s3 cp /backup/snapshot_backup.tar.gz s3://mybucket/"
+  else
+    aws s3 cp /backup/snapshot_backup.tar.gz s3://mybucket/
+  fi
 }
 
 main() {
   args_setup "$@"
-  if [ -e "$DEV_PATH" ] && echo "$S3_LINK" | grep "^s3://" ; then
-    log_message "Placeholder" "info"
-  else
-    log_message "Couldn't find valid partition file or s3 link. Check -p and -s options" "err"
-    exit 1
+  if [[ -z "$DEV_PATH" || -z "$S3_LINK" ]] ; then
+    help_info
   fi
   flush_cache
   create_snapshot
